+++ src/main/kotlin/com/yourssu/soongpt/domain/timetable/implement/LabTimetableMapper.kt
@@ -1,4 +1,4 @@
-package com.yourssu.soongpt.domain.timetable.business
+package com.yourssu.soongpt.domain.timetable.implement
 
 import com.yourssu.soongpt.domain.timetable.business.dto.*
 import com.yourssu.soongpt.domain.course.implement.Course
